rag-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
cohere>=5.0.0
qdrant-client>=1.12.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tiktoken>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
